[Fix] fix delete of 0 computers, button offset active class, company id sorting, and next page sorting type consistency

+++ computer_database/webapp/src/main/java/com/excilys/computer_database/controller/ComputerController.java
@@ -130,10 +130,15 @@ public class ComputerController extends ApplicationController{
         
         List<String> ids = Arrays.asList(request.getParameter("selection").split("\\s*,\\s*"));
 
-        for (String id : ids) {
-            computerService.delete(Integer.parseInt(id));
+        if(ids != null) {
+
+            for (String id : ids) {
+                if(id != null && !"".equals(id)) {
+                    computerService.delete(Integer.parseInt(id));
+                }
+            }
         }
-        
+
         return REDIRECT + JSP_DASHBOARD;
     }
 }
